'pre-beta'

+++ users/views.py
@@ -4,21 +4,19 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib.auth.models import User
 from django.forms import formset_factory
-from .forms import LoginForm, RegForm, UserInfoForm, TeamForm, TeamMemberForm, MemberFormset
-from .models import Teams, MemberTeam, Profile, Tasks
+from django.http import FileResponse
+from .forms import LoginForm, RegForm, UserInfoForm, TeamForm, TeamMemberForm, MemberFormset, HackathonForm, HackathonDateForm
+from .models import Teams, MemberTeam, Profile, Tasks, HackathonsRegulations, Hackathons
+
 
 # Create your views here.
 def profile(request):
-    if not User.objects.filter(username=request.user.username):
-        context = 'пользователь не существует'
-    else:
-        context = 'пользователь существует'
     if not request.user.is_authenticated:
         return HttpResponseRedirect(reverse("users:login"))
-    return render(request, "users/personal.html", {
-
-        'message':context,
-    })
+    elif request.user.is_superuser:
+        return render(request, "users/admin.html")
+    else:
+        return render(request, "users/personal.html")
 
 def login_view(request):
     if request.method == 'POST':
@@ -45,34 +43,41 @@ def logout_view(request):
     })
 
 def add_command(request):
-    teams = Teams.objects.all()
-    member_form = MemberFormset()
-    return render(request, "users/add_command.html", {
-        'teams': teams,
-        'formset':member_form
-    })
-
+    if request.user.is_authenticated:
+        teams = Teams.objects.all()
+        member_form = MemberFormset()
+        return render(request, "users/add_command.html", {
+            'teams': teams,
+            'formset':member_form
+        })
+    return render(request, "users/login.html")
 def del_command(request):
-    team = Teams.objects.filter(name=request.user.memberteam.team)
-    team.delete()
-    return HttpResponseRedirect(reverse("users:command"))
-
+    if request.user.is_authenticated:
+        team = Teams.objects.filter(name=request.user.memberteam.team)
+        team.delete()
+        return HttpResponseRedirect(reverse("users:command"))
+    return render(request, "users/login.html")
 def add_member(request):
-    member_form = TeamMemberForm(request.POST)
-    members = member_form.data.getlist('name')
-    users = User.objects.filter(username__in=members)
-    for user in users:
-        team_member = MemberTeam(team=request.user.memberteam.team, user=user)
-        team_member.save()
-    return HttpResponseRedirect(reverse("users:command"))
+    if request.user.is_authenticated:
+        member_form = TeamMemberForm(request.POST)
+        members = member_form.data.getlist('name')
+        users = User.objects.filter(username__in=members)
+        for user in users:
+            team_member = MemberTeam(team=request.user.memberteam.team, user=user)
+            team_member.save()
+        return HttpResponseRedirect(reverse("users:command"))
 
+    return render(request, "users/login.html")
 def del_member(request):
-    member_form = TeamMemberForm(request.POST)
-    members = member_form.data.getlist('name')
-    users = User.objects.filter(username__in=members)
-    for user in users:
-        user.memberteam.delete()
-    return HttpResponseRedirect(reverse("users:command"))
+    if request.user.is_authenticated:
+        member_form = TeamMemberForm(request.POST)
+        members = member_form.data.getlist('name')
+        users = User.objects.filter(username__in=members)
+        for user in users:
+            user.memberteam.delete()
+        return HttpResponseRedirect(reverse("users:command"))
+
+    return render(request, "users/login.html")
 def command(request):
 # Должна быть проверка на существование команды в БД -- ГОТОВО
 # Если существует, то запрет на создание -- ГОТОВО
@@ -80,43 +85,42 @@ def command(request):
 # Если все участники свободны, то создание команды и привязка участников к команде -- ГОТОВО
 # Если участник уже в команде, то вывод страницы с инфой о команде и настройками (в html можно настроить)
 # Проблема: Нужна таблица с командами, расширить модель User, обработка формы разных размеров -- ГОТОВО
-    if MemberTeam.objects.filter(user=request.user):
-        return render(request, "users/team.html", {'members':request.user.memberteam.team_members(team=request.user.memberteam.team)})
-    else:
+    if request.user.is_authenticated:
+        if MemberTeam.objects.filter(user=request.user):
+            return render(request, "users/team.html", {'members':request.user.memberteam.team_members(team=request.user.memberteam.team)})
+        else:
 
-        if request.method == 'POST':
-            team_form = TeamForm(request.POST)
-            if team_form.is_valid():
-                team_name = team_form.cleaned_data['team']
-                members = team_form.data.getlist('name')
-                users = User.objects.filter(username__in=members)
-                if (not Teams.objects.filter(name=team_name)) \
-                        and (users.count() == int(request.POST['form-TOTAL_FORMS']))\
-                        and not all(MemberTeam.objects.filter(user=user) for user in users):
-                    team = Teams(name=team_name)
-                    team.save()
-                    for user in users:
-                        team_member = MemberTeam(team=team, user=user)
+            if request.method == 'POST':
+                team_form = TeamForm(request.POST)
+                if team_form.is_valid():
+                    team_name = team_form.cleaned_data['team']
+                    members = team_form.data.getlist('name')
+                    users = User.objects.filter(username__in=members)
+                    if (not Teams.objects.filter(name=team_name)) \
+                            and (users.count() == int(request.POST['form-TOTAL_FORMS']))\
+                            and not all(MemberTeam.objects.filter(user=user) for user in users):
+                        team = Teams(name=team_name)
+                        team.save()
+                        for user in users:
+                            team_member = MemberTeam(team=team, user=user)
+                            team_member.save()
+                        team_member = MemberTeam(team=team, user=request.user)
                         team_member.save()
-                    team_member = MemberTeam(team=team, user=request.user)
-                    team_member.save()
-                    return render(request, "users/team.html",
-                                  {'members': request.user.memberteam.team_members(team=request.user.memberteam.team)})
+                        return render(request, "users/team.html",
+                                      {'members': request.user.memberteam.team_members(team=request.user.memberteam.team)})
 
-                else:
                     return render(request, "users/personal.html", {
                         'message':'Команда уже существует, либо пользователей которых вы пытаетесь добавить не существует, либо уже заняты в другой команде',
                     })
-
-                return HttpResponseRedirect(reverse("users:add_command"))
-            else:
 
                 return render(request, "users/personal.html", {
 
                     'message': 'форма/не/валидна',
                 })
-        else:
+
             return HttpResponseRedirect(reverse("users:profile"))
+
+    return render(request, "users/login.html")
 def register(request):
 
     if request.method == 'POST':
@@ -137,11 +141,11 @@ def register(request):
                 return render(request, "users/registration.html", {
                     'message': 'Пользователь существует',
                 })
-
-    return render(request, "users/registration.html", {
-        'message':'Invalid form',
-    })
-
+        else:
+            return render(request, "users/registration.html", {
+                'message':'Неверная форма',
+            })
+    return render(request, "users/registration.html")
 
 
 def update_user_info(request):
@@ -200,33 +204,104 @@ def update_user_info(request):
         return HttpResponseRedirect(reverse("users:profile"))
 
 def delete_user(request):
-    user = User.objects.get(username=request.user.username)
-    User.delete(user)
-    logout(request)
-    return render(request, 'users/login.html', {
-        'message': 'Вы вышли из системы',
-    })
-
-def schedule(request):
-    return render(request, 'users/schedule.html',)
-
-def task(request):
-    if request.user.memberteam.team.task_id:
-        return render(request, 'users/user_task.html', {
-            'link':request.user.memberteam.team.task_id.link_to_the_task_text
+    if request.user.is_authenticated:
+        user = User.objects.get(username=request.user.username)
+        User.delete(user)
+        logout(request)
+        return render(request, 'users/login.html', {
+            'message': 'Вы вышли из системы',
         })
-    else:
-        return render(request, 'users/tasks.html', )
+    return render(request, "users/login.html")
+def schedule(request):
+    if request.user.is_authenticated:
+        return render(request, 'users/schedule.html',{
+            'timetable':HackathonsRegulations.objects.all()
+        })
+    return render(request, "users/login.html")
+def task(request):
+    if request.user.is_authenticated:
+        if MemberTeam.objects.filter(user=request.user):
+            if request.user.memberteam.team.task_id:
+                return render(request, 'users/user_task.html', {
+                    'link':request.user.memberteam.team.task_id.link_to_the_task_text
+                })
+        else:
+            return render(request, 'users/tasks.html', )
+    return render(request, "users/login.html")
 
 def new_task(request):
-    team = request.user.memberteam.team
-    task = Tasks.objects.get(link_to_the_task_text=request.POST['link'])
-    team.task_id = task
-    team.save()
-    return HttpResponseRedirect(reverse('users:task'))
+    if request.user.is_authenticated:
+        if MemberTeam.objects.filter(user=request.user):
+            team = request.user.memberteam.team
+            task = Tasks.objects.get(link_to_the_task_text=request.POST['link'])
+            team.task_id = task
+            team.save()
+            return HttpResponseRedirect(reverse('users:task'))
+        return render(request, "users/personal.html", {
+            'message': 'Соберите команду, чтобы выбрать команду',
+        })
+    return render(request, "users/login.html")
 
 def change_task(request):
-    team = request.user.memberteam.team
-    team.task_id = None
-    team.save()
-    return HttpResponseRedirect(reverse('users:task'))
+    if request.user.is_authenticated:
+        team = request.user.memberteam.team
+        team.task_id = None
+        team.save()
+        return HttpResponseRedirect(reverse('users:task'))
+    return render(request, "users/login.html")
+#Admin
+def new_hackathon(request):
+    if request.user.is_authenticated and request.user.is_superuser:
+        if not Hackathons.objects.all() and request.method == 'POST':
+            form = HackathonForm(request.POST)
+
+            if form.is_valid():
+                cd = form.cleaned_data
+
+                hackathon = Hackathons(name=cd['name'],
+                                        start_date=cd['start_date'],
+                                        end_date=cd['end_date'], )
+
+                hackathon.save()
+        return render(request, "users/admin.html", {
+            'message': 'Не может быть более двух Хакатонов одновременно',
+        })
+    return render(request, "users/login.html")
+
+def update_date_hackathon(request):
+    if request.user.is_authenticated and request.user.is_superuser:
+        if request.method == 'POST':
+            form = HackathonDateForm(request.POST)
+
+            if form.is_valid():
+                cd = form.cleaned_data
+                hackathon = Hackathons.objects.all()[0]
+                hackathon.start_date = cd['start_date']
+                hackathon.end_date = cd['end_date']
+
+                hackathon.save()
+            return HttpResponseRedirect(reverse('hackaton:index'))
+        return render(request, "users/admin.html", {
+            'message': 'Произошла ошибка',
+        })
+    return render(request, "users/login.html")
+
+
+#ДОДЕЛАТЬ!
+def add_representative(request):
+    if request.user.is_authenticated and request.user.is_superuser:
+        if not Hackathons.objects.all() and request.method == 'POST':
+            form = HackathonForm(request.POST)
+
+            if form.is_valid():
+                cd = form.cleaned_data
+
+                hackathon = Hackathons(name=cd['name'],
+                                        start_date=cd['start_date'],
+                                        end_date=cd['end_date'], )
+
+                hackathon.save()
+        return render(request, "users/admin.html", {
+            'message': 'Не может быть более двух Хакатонов одновременно',
+        })
+    return render(request, "users/login.html")
